change download option, add backend, update readme
